// src/app/Models/Addition.ts
export class Addition {
  addition: number = 0;
  prodid: string = '';
  name: string = '';
  IsMeal?: number | boolean | null;
  IsFirstMeal?: number | boolean| null;
  NumberOfChoices?: number;
  Image?: string;
  IsFavorite?: number| boolean| null;
  IsAvaliable?: number| boolean | null;
  Style?: string;
  Checked?: number| boolean| null;
  IsSauce?: number| boolean| null;
  isLastMeal?: number| boolean| null;
  isBread?: number| boolean| null;
  extraAddition?: number| boolean| null;
  additionMessage?: any;
  haveException?: any;
  isPowerBread?: string;
  AdditionBlock?: string | null;
  IsTop?: string | boolean;
  new?: boolean;
}